requirements.txt
fastapi==0.68.2
uvicorn==0.17.4
motor==2.5.1
pytest==7.1.1
email-validator==1.1.3
starlette==0.14.2
python-decouple==3.6